// apps/auth/src/constants/auth.enum.ts
export enum ENUM_USER_SIGN_UP_FROM {
  LOCAL = 'LOCAL',
  PUBLIC = 'PUBLIC',
  ADMIN = 'ADMIN',
}
export enum ENUM_USER_STATUS_CODE_ERROR {
  USER_NOT_FOUND_ERROR = 5200,
  USER_USERNAME_EXISTS_ERROR = 5201,
  USER_EMAIL_EXIST_ERROR = 5202,
  USER_MOBILE_NUMBER_EXIST_ERROR = 5203,
  USER_IS_ACTIVE_ERROR = 5204,
  USER_INACTIVE_ERROR = 5205,
  USER_INACTIVE_PERMANENT_ERROR = 5206,
  USER_PASSWORD_NOT_MATCH_ERROR = 5207,
  USER_PASSWORD_NEW_MUST_DIFFERENCE_ERROR = 5208,
  USER_PASSWORD_EXPIRED_ERROR = 5209,
  USER_PASSWORD_ATTEMPT_MAX_ERROR = 5210,
  USER_BLOCKED_ERROR = 5211,
  USER_IS_NOT_VERIFY = 403,
}
export enum UserVerifyStatus {
  Unverified,
  Verified,
}
export enum ENUM_ROLE_STATUS_CODE_ERROR {
  ROLE_NOT_FOUND_ERROR = 5100,
  ROLE_PAYLOAD_TYPE_INVALID_ERROR = 5101,
  ROLE_EXIST_ERROR = 5102,
  ROLE_IS_ACTIVE_ERROR = 5103,
  ROLE_INACTIVE_ERROR = 5104,
  ROLE_USED_ERROR = 5105,
}